chore: Address Obsidian plugin feedback (#44)
PR to obsidian releases:
https://github.com/obsidianmd/obsidian-releases/pull/6279

Fixing the following issues:
1.
[[1](https://github.com/LeonLuttenberger/obsidian-kindle-highlight-import/blob/b9675fde50fabb1ed81dd57ceb1ca4c377e953e2/main.ts#L47-L109)]:Remember
to rename the sample classes to something that makes sense to your
plugin. If there are any errors, Obsidian will spit out this class name,
so this will be important for users to debug if they need to.
2.
[[1](https://github.com/LeonLuttenberger/obsidian-kindle-highlight-import/blob/b9675fde50fabb1ed81dd57ceb1ca4c377e953e2/src/components/fileUploadModal.ts#L21)]:You
should avoid assigning styles via JavaScript or in HTML and instead move
all these styles into CSS so that they are more easily adaptable by
themes and snippets.
3.
[[1](https://github.com/LeonLuttenberger/obsidian-kindle-highlight-import/blob/b9675fde50fabb1ed81dd57ceb1ca4c377e953e2/src/components/fileUploadModal.ts#L23)]:Using
innerHTML, outerHTML or similar API's is a security risk. Instead, use
the DOM API or the Obsidian helper functions:
https://docs.obsidian.md/Plugins/User+interface/HTML+elements